+++ lib/features/hooks/hooks.rb
@@ -4,7 +4,7 @@ require 'cucumber'
 require 'json'
 require 'securerandom'
 
-AfterConfiguration do |config|
+AfterConfiguration do |_cucumber_config|
 
   # Start mock server
   Server.start_server
@@ -18,57 +18,90 @@ AfterConfiguration do |config|
   if config.farm == :bs
     tunnel_id = SecureRandom.uuid
     config.capabilities = Capabilities.for_browser_stack config.bs_device,
-                                                         tunnel_id
+                                                         tunnel_id,
+                                                         config.appium_version,
+                                                         config.capabilities_option
 
-    config.app_location = BrowserStackUtils.upload_app config.username,
-                                                       config.access_key,
-                                                       config.app_location
+    config.app = BrowserStackUtils.upload_app config.username,
+                                              config.access_key,
+                                              config.app
     BrowserStackUtils.start_local_tunnel config.bs_local,
                                          tunnel_id,
                                          config.access_key
   elsif config.farm == :local
     config.capabilities = Capabilities.for_local config.os,
+                                                 config.capabilities_option,
                                                  config.apple_team_id,
                                                  config.device_id
   end
-  # Set app location (file or url) in capabilities
-  config.capabilities['app'] = config.app_location
 
-  # Create and start the driver
-  MazeRunner.driver = ResilientAppiumDriver.new config.appium_server_url,
-                                                config.capabilities,
-                                                config.locator
+  # Set app location (file or url) in capabilities
+  config.capabilities['app'] = config.app
+
+  # Create and start the relevant driver
+  MazeRunner.driver = if MazeRunner.config.resilient
+                        $logger.info 'Creating ResilientAppiumDriver instance'
+                        ResilientAppiumDriver.new config.appium_server_url,
+                                                  config.capabilities,
+                                                  config.locator
+                      else
+                        $logger.info 'Creating AppiumDriver instance'
+                        AppiumDriver.new config.appium_server_url,
+                                         config.capabilities,
+                                         config.locator
+                      end
   if config.farm == :bs
     # Log a link to the BrowserStack session search dashboard
     build = MazeRunner.driver.caps[:build]
     url = "https://app-automate.browserstack.com/dashboard/v2/?searchQuery=#{build}"
-    $logger.info LogUtil.linkify url, 'BrowserStack session(s)'
+    if ENV['BUILDKITE']
+      $logger.info LogUtil.linkify url, 'BrowserStack session(s)'
+    else
+      $logger.info "BrowserStack session(s): #{url}"
+    end
   end
   MazeRunner.driver.start_driver unless config.appium_session_isolation
+
+  # Call any blocks registered by the client
+  MazeRunner.hooks.call_after_configuration config
 end
 
 # Before each scenario
 Before do |scenario|
   STDOUT.puts "--- Scenario: #{scenario.name}"
-  # TODO: PLAT-5309 - Building a clear environment for each scenario into MazeRunner would be a good
-  #   thing, but a number of our pipelines rely on certain variables existing throughout (e.g. Ruby).
-  # Runner.environment.clear
+
+  Runner.environment.clear
   Server.stored_requests.clear
+  Server.invalid_requests.clear
   Store.values.clear
 
-  next if MazeRunner.config.farm == :none
+  MazeRunner.driver.start_driver if MazeRunner.config.farm != :none && MazeRunner.config.appium_session_isolation
 
-  MazeRunner.driver.start_driver if MazeRunner.config.appium_session_isolation
+  # Launch the app on macOS
+  MazeRunner.driver.get(MazeRunner.config.app) if MazeRunner.config.os == 'macos'
+
+  # Call any blocks registered by the client
+  MazeRunner.hooks.call_before scenario
 end
 
-# After each scenario
+# General processing to be run after each scenario
 After do |scenario|
+
+  # Call any blocks registered by the client
+  MazeRunner.hooks.call_after scenario
+
+  # Make sure we reset to HTTP 200 return status after each scenario
+  Server.status_code = 200
+  Server.reset_status_code = false
 
   # This is here to stop sessions from one test hitting another.
   # However this does mean that tests take longer.
+  # In addition, reset the last captured exit code
   # TODO:SM We could try and fix this by generating unique endpoints
   # for each test.
   Docker.down_all_services
+  Docker.last_exit_code = nil
+  Docker.last_command_logs = nil
 
   # Make sure that any scripts are killed between test runs
   # so future tests are run from a clean slate.
@@ -80,7 +113,7 @@ After do |scenario|
   if scenario.failed?
     STDOUT.puts '^^^ +++'
     if Server.stored_requests.empty?
-      $logger.info 'No requests received'
+      $logger.info 'No valid requests received'
     else
       $logger.info 'The following requests were received:'
       Server.stored_requests.each.with_index(1) do |request, number|
@@ -94,13 +127,34 @@ After do |scenario|
 
   if MazeRunner.config.appium_session_isolation
     MazeRunner.driver.driver_quit
+  elsif MazeRunner.config.os == 'macos'
+    # Close the app - without the sleep, launching the app for the next scenario intermittently fails
+    system("killall #{MazeRunner.config.app} && sleep 1")
   else
     MazeRunner.driver.reset_with_timeout 2
+  end
+
+  STDOUT.puts scenario.inspect
+end
+
+# Check for invalid requests after each scenario.  In its own hook as failing a scenario raises an exception.
+# Furthermore, this hook should appear after the general hook as they are exectued in reverse order by Cucumber.
+After do |scenario|
+  unless Server.invalid_requests.empty?
+    Server.invalid_requests.each.with_index(1) do |request, number|
+      $logger.error "Invalid request #{number} (#{request[:reason]}):"
+      LogUtil.log_hash(Logger::Severity::ERROR, request)
+    end
+    msg = "#{Server.invalid_requests.length} invalid request(s) received during scenario"
+    scenario.fail msg
   end
 end
 
 # After all tests
 at_exit do
+
+  STDOUT.puts '+++ All scenarios complete'
+
   # Stop the mock server
   Server.stop_server
 
@@ -114,4 +168,5 @@ at_exit do
   # Stop the Appium session
   MazeRunner.driver.driver_quit unless MazeRunner.config.appium_session_isolation
 end
+