+++ lib/features/support/maze_runner.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative './configuration'
+require_relative '../hooks/client_hooks'
 
 # Glues the various parts of MazeRunner together that need to be accessed globally,
 # providing an alternative to the proliferation of global variables or singletons.
@@ -8,7 +9,11 @@ class MazeRunner
   class << self
     attr_accessor :driver
     def config
-      @config ||= Configuration.new
+      @config ||= Maze::Configuration.new
+    end
+
+    def hooks
+      @hooks ||= Hooks.new
     end
   end
 end