+++ lib/features/support/runner.rb
@@ -1,8 +1,10 @@
+# frozen_string_literal: true
+
 require 'open3'
 require_relative './interactive_cli'
 
 # Determines the default path to the `scripts` directory.  Can be overwritten in the env.rb file.
-SCRIPT_PATH ||= File.expand_path(File.join(File.dirname(__FILE__), "..", "scripts")) unless defined? SCRIPT_PATH
+SCRIPT_PATH ||= File.expand_path(File.join(File.dirname(__FILE__), '..', 'scripts')) unless defined? SCRIPT_PATH
 
 # Runs scripts and commands, applying relevant environment variables as necessary
 class Runner
@@ -15,27 +17,27 @@ class Runner
     # @param blocking [Boolean] Optional. Whether to wait for a return code before proceeding
     # @param success_codes [Array] Optional. An array of integer codes which indicate the run was successful
     #
-    # @return [Array] If blocking, the output and exit_status are returned
+    # @return [Array, Thread] If blocking, the output and exit_status are returned
     def run_command(cmd, blocking: true, success_codes: [0])
       executor = lambda do
-        $logger.debug(cmd) { 'executing' }
+        $logger.debug "Executing: #{cmd}"
 
-        Open3.popen2e(environment, cmd) do |stdin, stdout_and_stderr, wait_thr|
+        Open3.popen2e(environment, cmd) do |_stdin, stdout_and_stderr, wait_thr|
           # Add the pid to the list of pids to kill at the end
           pids << wait_thr.pid unless blocking
 
           output = []
           stdout_and_stderr.each do |line|
             output << line
-            $logger.debug(cmd) {line}
+            $logger.debug line
           end
 
           exit_status = wait_thr.value.to_i
-          $logger.debug(cmd) { "exit status: #{exit_status}" }
+          $logger.debug "Exit status: #{exit_status}"
 
           # if the command fails we log the output at warn level too
-          if success_codes != nil && !success_codes.include?(exit_status) && $logger.level != Logger::DEBUG
-            output.each {|line| $logger.warn(cmd) {line}}
+          if !success_codes.nil? && !success_codes.include?(exit_status) && $logger.level != Logger::DEBUG
+            output.each { |line| $logger.warn(cmd) { line } }
           end
 
           return [output, exit_status]
@@ -45,7 +47,7 @@ class Runner
       if blocking
         executor.call
       else
-        Thread.new &executor
+        Thread.new(&executor)
       end
     end
 
@@ -67,7 +69,7 @@ class Runner
         # using ruby etc.
         script_path = "cmd /c #{script_path}"
       end
-      return run_command(script_path, blocking: blocking, success_codes: success_codes)
+      run_command(script_path, blocking: blocking, success_codes: success_codes)
     end
 
     # Returns the current interactive session, creating a new one if necessary
@@ -78,6 +80,7 @@ class Runner
       attempts = 0
       until @interactive_session.pid
         raise "Shell session would not start within 3 seconds" if attempts > 10 # allows 3 seconds of opening
+
         attempts += 1
         sleep(0.3)
       end
@@ -87,6 +90,7 @@ class Runner
     # Stops the interactive session, allowing a new one to be started
     def stop_interactive_session
       return unless @interactive_session&.running
+
       @interactive_session.stop
       # Make sure the process is properly ended
       pids << @interactive_session.pid
@@ -96,12 +100,11 @@ class Runner
     # Stops all script processes previously started by this class.
     def kill_running_scripts
       stop_interactive_session
-      pids.each {|p|
-        begin
-          Process.kill("KILL", p)
-        rescue Errno::ESRCH
-        end
-      }
+      pids.each do |p|
+        Process.kill('KILL', p)
+      rescue Errno::ESRCH
+        # ignored
+      end
       pids.clear
     end
 
@@ -109,7 +112,7 @@ class Runner
     #
     # @return [Hash] The hash of currently set environment variables and their values
     def environment
-      @env ||= {}
+      @environment ||= {}
     end
 
     private